chore(virtctl): Move guest os info command to own file
